chore: storybook preview 설정 변경
